+++ find/src/main/java/com/brins/find/presenter/FindPresenter.kt
@@ -8,7 +8,7 @@ import com.brins.find.contract.FindContract
  */
 class FindPresenter : FindContract.Presenter() {
 
-    override suspend fun loadEvent(lastTime: Int, pageSize: Int) {
+    override suspend fun loadEvent(lastTime: Long, pageSize: Int) {
         val result = mModel?.loadUserEvent(lastTime, pageSize)
         result?.let {
             mView?.onUserEventLoad(it)
